SMS (Short Message Service)
Today, when writing SMS (Short Message Service) messages, the messages are shortened by removing the vowels in the words due to the 160 character limit and the effort to write fast. If the first letter of the word is a vowel, it remains.Write the function called makeShortSMS, which turns a sentence entered into the keyboard into a shortened SMS message, together with a main function where the original of the message is entered from the keyboard and the abbreviated version is printed on the screen. The function copies the shortened version of the original message received in the first parameter to the second parameter.